ternary and string support, if_no_else made smaller

+++ stress_testing/74408ee0-ce05-49d0-b539-79efb911d2fd.c
@@ -2,5 +2,5 @@
 
 int main() {
     int a = 10, b = 3;
-    float max = (a > b) ? a : b;
+    int max = (a > b) ? a : b;
 }
